move: 카테고리 폴더구조 변경
