+++ src/main/java/com/homefit/backend/exercise/service/ExerciseService.java
@@ -4,9 +4,9 @@ import com.homefit.backend.exercise.dto.ExerciseRequest;
 import com.homefit.backend.exercise.dto.ExerciseResponse;
 import com.homefit.backend.exercise.entity.Exercise;
 import com.homefit.backend.exercise.repository.ExerciseRepository;
-import com.homefit.backend.exerciseCategory.entity.ExerciseCategory;
-import com.homefit.backend.exerciseCategory.service.ExerciseCategoryService;
-import com.homefit.backend.exerciseLog.service.ExerciseLogService;
+import com.homefit.backend.category.exercise.entity.ExerciseCategory;
+import com.homefit.backend.category.exercise.service.ExerciseCategoryService;
+import com.homefit.backend.exerciselog.service.ExerciseLogService;
 import com.homefit.backend.global.exception.ErrorCode;
 import com.homefit.backend.global.exception.model.ConflictException;
 import com.homefit.backend.global.exception.model.NotFoundException;
