--- src/main/java/com/homefit/backend/exerciseCategory/dto/ExerciseCategoryRequest.java
@@ -1,16 +1,0 @@
-package com.homefit.backend.exerciseCategory.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.*;
-
-@Getter
-@ToString
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Schema(description = "운동 카테고리 생성 요청 DTO")
-public class ExerciseCategoryRequest {
-
-    @Schema(description = "운동 카테고리 이름", example = "맨몸운동")
-    private String name;
-}
