+++ src/main/java/com/homefit/backend/item/dto/ItemModifyRequestDto.java
@@ -1,6 +1,6 @@
 package com.homefit.backend.item.dto;
 
-import com.homefit.backend.itemcategory.entity.ItemCategory;
+import com.homefit.backend.category.item.entity.ItemCategory;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
