+++ src/main/java/com/homefit/backend/item/dto/ItemResponseDto.java
@@ -1,9 +1,8 @@
 package com.homefit.backend.item.dto;
 
-import com.homefit.backend.itemcategory.entity.ItemCategory;
+import com.homefit.backend.category.item.entity.ItemCategory;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Data
 @NoArgsConstructor
