+++ src/main/java/com/homefit/backend/item/entity/Item.java
@@ -1,7 +1,7 @@
 package com.homefit.backend.item.entity;
 
 import com.homefit.backend.item.dto.ItemModifyRequestDto;
-import com.homefit.backend.itemcategory.entity.ItemCategory;
+import com.homefit.backend.category.item.entity.ItemCategory;
 import jakarta.persistence.*;
 import lombok.*;
