+++ src/test/java/com/homefit/backend/exercise/service/ExerciseServiceTest.java
@@ -4,9 +4,9 @@ import com.homefit.backend.exercise.dto.ExerciseRequest;
 import com.homefit.backend.exercise.dto.ExerciseResponse;
 import com.homefit.backend.exercise.entity.Exercise;
 import com.homefit.backend.exercise.repository.ExerciseRepository;
-import com.homefit.backend.exerciseCategory.entity.ExerciseCategory;
-import com.homefit.backend.exerciseCategory.repository.ExerciseCategoryRepository;
-import com.homefit.backend.exerciseLog.service.ExerciseLogService;
+import com.homefit.backend.category.exercise.entity.ExerciseCategory;
+import com.homefit.backend.category.exercise.repository.ExerciseCategoryRepository;
+import com.homefit.backend.exerciselog.service.ExerciseLogService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
